LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/browsers/sandboxing/sandbox-disallow-same-origin-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://web-platform.test:8800/html/browsers/sandboxing/inner-iframe.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Access to sandbox iframe


PASS Access to sandbox iframe is disallowed
